Refactoring functions and adding comments

+++ RelayTestApp/Plugins/BCClient/Source/BCClientPlugin/Public/IRelayCallback.h
@@ -7,5 +7,5 @@ class BCCLIENTPLUGIN_API IRelayCallback
   public:
     /**
      */
-    virtual void relayCallback(uint8 netId,const TArray<uint8>& jsonData) = 0;
+    virtual void relayCallback(int netId,const TArray<uint8>& jsonData) = 0;
 };
